# perf(lsp): request only changed portions of the buffer in changetracker