I just updated https://github.com/neovim/neovim/pull/16277 to address the conflicts. @dmitmel is the original author of this PR (feel free to just take these changes and apply it to the other PR)

The following code measures the time required to perform a buffer edit to
that operates individually on each line, common to plugins such as vim-commentary.

    set rtp+=~/.config/nvim/plugged/nvim-lspconfig
    set rtp+=~/.config/nvim/plugged/vim-commentary
    lua require('lspconfig')['ccls'].setup({})
    function! Benchmark(tries) abort
        let results_comment = []
        let results_undo = []
        for i in range(a:tries)
            echo printf('run %d', i+1)
            let begin = reltime()
            normal gggcG
            call add(results_comment, reltimefloat(reltime(begin)))
            let begin = reltime()
            silent! undo
            call add(results_undo, reltimefloat(reltime(begin)))
            redraw
        endfor
        let avg_comment = 0.0
        let avg_undo = 0.0
        for i in range(a:tries)
            echomsg printf('run %3d: comment=%fs undo=%fs', i+1, results_comment[i], results_undo[i])
            let avg_comment += results_comment[i]
            let avg_undo += results_undo[i]
        endfor
        echomsg printf('average: comment=%fs undo=%fs', avg_comment / a:tries, avg_undo / a:tries)
    endfunction
    command! -bar Benchmark call Benchmark(10)

All text changes will be recorded within a single undo operation. Both the
comment operation itself and the undo operation will generate an on_lines event
for each changed line. Formatter plugins using setline() have also been found
to exhibit the same problem (neoformat, :RustFmt in rust.vim), as this function
too generates an on_lines event for every line it changes.

Using the neovim codebase as an example (commit 2ecf0a4)
with neovim itself built at 2ecf0a4 with
CMAKE_BUILD_TYPE=Release shows the following performance improvement:

src/nvim/lua/executor.c, 1432 lines:
  baseline, no optimizations:             comment=0.540587s undo=0.440249s
  without double-buffering optimization:  comment=0.183314s undo=0.060663s
  all optimizations in this commit:       comment=0.174850s undo=0.052789s

src/nvim/search.c, 5467 lines:
  baseline, no optimizations:             comment=7.420446s undo=7.656624s
  without double-buffering optimization:  comment=0.889048s undo=0.486026s
  all optimizations in this commit:       comment=0.662899s undo=0.243628s

src/nvim/eval.c, 11355 lines:
  baseline, no optimizations:             comment=41.775695s undo=44.583374s
  without double-buffering optimization:  comment=3.643933s undo=2.817158s
  all optimizations in this commit:       comment=1.510886s undo=0.707928s

This performance improvement is due to avoiding expensive lua string reallocations
on each operation, and instead requesting only the changed chunk of the buffer,
as reported by firstline and new_lastline parameters of on_lines, plus re-using
already allocated tables for storing the resulting lines to reduce the load on
the garbage collector.

Co-authored-by: Dmytro Meleshko <dmytro.meleshko@gmail.com>